tried to delete stack - FAIL

+++ include/picell.h
@@ -23,7 +23,7 @@ enum {
 
 typedef struct cell {
   unsigned char type, marked;
-  unsigned short marks;
+  unsigned long marks;
   union {
     struct {
       struct cell *car;
